+++ src/main/java/com/trustamarket/walletservice/infrastructure/persistence/PointTransactionRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.trustamarket.walletservice.infrastructure;
+package com.trustamarket.walletservice.infrastructure.persistence;
 
 import java.util.List;
 
@@ -6,6 +6,7 @@ import org.springframework.stereotype.Repository;
 
 import com.trustamarket.walletservice.domain.entity.PointTransaction;
 import com.trustamarket.walletservice.domain.repository.PointTransactionRepository;
+import com.trustamarket.walletservice.infrastructure.persistence.jpa.PointTransactionJpaRepository;
 
 import lombok.RequiredArgsConstructor;
